Shift plexus to TypeScript (from flowtypes) (#331)
* Shift plexus to TypeScript (from flowtypes)

Main outstanding items are to add linting to plexus TS files and more
comments in the plexus build related files.

Signed-off-by: Joe Farro <joef@uber.com>

* Typo from renaming

Signed-off-by: Joe Farro <joef@uber.com>

* Add linting for plexus TypeScript

This required an update to `eslint-plugin-import`, which, in turn,
emtailed updating a few packages. This resulted in the addition of a
few new rules. So, there are also changes to adhere to the new rules.

Signed-off-by: Joe Farro <joef@uber.com>

* Fix a few issues from merging master

Signed-off-by: Joe Farro <joef@uber.com>

* Better linting for plexus / TypeScript

TypeScript linting is not quite correctly intergrated with VS code, but
the current eslint and tsc-lint npm scripts have TS linting in pretty
good shape. tsc-lint endeavors to build the TS files but does not emit
them, thereby filling in the gaps left by eslint TS support.

Signed-off-by: Joe Farro <joef@uber.com>

* Fix lint error from merging master

Signed-off-by: Joe Farro <joef@uber.com>

* TypeScript type cleanup

Signed-off-by: Joe Farro <joef@uber.com>

* Build docs for root and plexus, revive plexus demo

- Main type exports moved from src/types/layout to src/types/index
- Change worker alias to be more obvious, e.g. "worker-alias"
- Various props made optional on:
  - DirectedGraph
  - PureEdges
  - PureNodes

Signed-off-by: Joe Farro <joef@uber.com>

* Update changelog

Signed-off-by: Joe Farro <joef@uber.com>

* Misc cleanup and add GitHub tickets to build docs

Signed-off-by: Joe Farro <joef@uber.com>

* Add license to package.json for publishing

Signed-off-by: Joe Farro <joef@uber.com>